apperently theres no need to put the febbox base into the env; Also refactored it a bit so there are seperate source lists for movies and for shows; I barely care that im gonna repeat like 50% of my code now!!
